Рефакторинг экранов настроек с использованием Voyager
Этот коммит рефакторит экраны настроек, заменяя Activity на Screen из библиотеки Voyager для навигации.

**Основные изменения:**

-   **Удалены Activity:**
    -   `AppearanceActivity.kt`
    -   `NotifyActivity.kt`
    -   `TestScreen.kt` (пустой файл)
-   **Переименован и отрефакторен `AboutActivity.kt` в `AboutScreen.kt`:**
    -   Теперь наследуется от `cafe.adriel.voyager.core.screen.Screen`.
    -   Использует `LocalNavigator.currentOrThrow` для навигации назад.
    -   Улучшено отображение информации о версии с использованием `AppInfo.getVersionName()`.
    -   Используется `SettingsItem` для элементов списка.
    -   Индикатор загрузки заменен на `CircularProgressIndicator()`.
-   **Создан `AppearanceScreen.kt`:**
    -   Новый экран, реализующий логику удаленного `AppearanceActivity.kt`.
    -   Наследуется от `cafe.adriel.voyager.core.screen.Screen`.
    -   Использует `LocalNavigator.currentOrThrow` для навигации.
    -   Улучшена логика сохранения и применения темы.
-   **Создан `NotificationsScreen.kt`:**
    -   Новый экран, заменяющий удаленный `NotifyActivity.kt`.
    -   Наследуется от `cafe.adriel.voyager.core.screen.Screen`.
    -   Использует `LocalNavigator.currentOrThrow` для навигации.
    -   Пока содержит базовую структуру Scaffold.
-   **Обновлен `SettingsScreen.kt`:**
    -   Теперь наследуется от `cafe.adriel.voyager.core.screen.Screen`.
    -   `MenuItem` теперь содержит лямбда-выражение `navigator: () -> Unit?` вместо `activityClass: Class<*>`.
    -   Использует `LocalNavigator.current` для навигации на дочерние экраны.
    -   Обновлено отображение версии приложения с использованием `AppInfo.getAppName()` и `AppInfo.getVersionName()`.
-   **Обновлен `StatisticsScreen.kt`:**
    -   Теперь наследуется от `cafe.adriel.voyager.core.screen.Screen`.
    -   Удалена неиспользуемая логика, связанная с `ChallengeViewModel`, `UserPrefs` и отображением аватара пользователя (эта функциональность, вероятно, будет перенесена или реализована иначе).
    -   Упрощен заголовок TopAppBar.

**Цель рефакторинга:**

-   Упрощение навигации в приложении за счет использования единого подхода с Voyager.
-   Улучшение структуры кода и разделение ответственности между экранами.
-   Подготовка к дальнейшему развитию функциональности настроек.
