--- app/src/main/java/com/wheatley/morph/layouts/settings/TestScreen.kt
@@ -1,13 +1,0 @@
-package com.wheatley.morph.layouts.settings
-
-import androidx.compose.runtime.Composable
-import cafe.adriel.voyager.core.screen.Screen
-
-class TestScreen: Screen {
-
-    @Composable
-    override fun Content() {
-        TODO("Not yet implemented")
-    }
-
-}
